requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
executing==1.2.0
Flask==2.2.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.5.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
lightgbm==3.3.5
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
openturns==1.20.post3
packaging==23.1
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post4
SQLAlchemy==2.0.15
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
waitress==2.1.2
wcwidth==0.2.6
Werkzeug==2.2.3
xgboost==1.7.5
zipp==3.15.0
